refactor [#7] 구조 리팩토링 및 api 키 설정 변경
