+++ src/main/java/cotato/five/weather/adapter/in/web/WeatherController.java
@@ -3,10 +3,9 @@ package cotato.five.weather.adapter.in.web;
 import cotato.five.weather.application.dto.WeatherDailyResponse;
 import cotato.five.weather.application.dto.WeatherHourlyResponse;
 import cotato.five.weather.application.dto.WeatherWeeklyResponse;
-import cotato.five.weather.application.service.WeatherService;
+import cotato.five.weather.application.WeatherService;
 import cotato.five.weather.common.ApiResponse;
 import cotato.five.weather.common.BaseResponse;
-import cotato.five.weather.common.SuccessResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.ResponseEntity;
