--- src/main/resources/application.properties
@@ -1,1 +1,0 @@
-openweather.api.key=2cad1ffa7b8e3e31918a57701cb33774
